Use FilteredConnection for campaigns list
