+++ web/src/enterprise/campaigns/global/list/GlobalCampaignListPage.tsx
@@ -1,28 +1,37 @@
 import React from 'react'
-import { CampaignList } from '../../list/CampaignList'
-import { useCampaigns } from '../../list/useCampaigns'
+import { queryCampaigns } from './backend'
 import AddIcon from 'mdi-react/AddIcon'
 import { Link } from '../../../../../../shared/src/components/Link'
+import { RouteComponentProps } from 'react-router'
+import { FilteredConnection } from '../../../../components/FilteredConnection'
+import { ICampaign } from '../../../../../../shared/src/graphql/schema'
+import { CampaignNode } from '../../list/CampaignNode'
 
-interface Props {}
+interface Props extends Pick<RouteComponentProps, 'history' | 'location'> {}
 
 /**
  * A list of all campaigns on the Sourcegraph instance.
  */
 export const GlobalCampaignListPage: React.FunctionComponent<Props> = props => {
-    const campaigns = useCampaigns()
     return (
         <>
             <h1>Campaigns</h1>
             <p>Track large-scale code changes</p>
-            {campaigns && campaigns.nodes.length > 0 && (
-                <div className="text-right mb-1">
-                    <Link to="/campaigns/new" className="btn btn-primary">
-                        <AddIcon className="icon-inline" /> New Campaign
-                    </Link>
-                </div>
-            )}
-            <CampaignList {...props} campaigns={campaigns} />
+
+            <div className="text-right mb-1">
+                <Link to="/campaigns/new" className="btn btn-primary">
+                    <AddIcon className="icon-inline" /> New Campaign
+                </Link>
+            </div>
+
+            <FilteredConnection<ICampaign>
+                {...props}
+                nodeComponent={CampaignNode}
+                queryConnection={queryCampaigns}
+                hideSearch={true}
+                noun="Campaign"
+                pluralNoun="Campaigns"
+            />
         </>
     )
 }
